refactor(webview): share the injected JS bridge across platforms
BuildInitScript — the ~100-line JS proxy/bridge injected into every page —
was byte-identical in webview_macos.mm, webview_linux.cc and
webview_windows.cc. Move it to webview/src/init_script.h; each platform file
includes it and passes only its host-specific postMessage expression.

Removes ~200 lines of duplication.
