+++ webview/src/init_script.h
@@ -1,0 +1,118 @@
+// Copyright 2025 Divy Srivastava. All rights reserved. MIT license.
+
+// The JavaScript bridge injected into every page is identical across all
+// platform webview backends; only the host-specific postMessage expression and
+// the namespace differ. Keep one copy here instead of duplicating it in
+// webview_macos.mm / webview_linux.cc / webview_windows.cc.
+
+#ifndef WEF_WEBVIEW_INIT_SCRIPT_H_
+#define WEF_WEBVIEW_INIT_SCRIPT_H_
+
+#include <string>
+
+// Builds the page init script. `ns` is the global namespace the proxy is
+// installed under (e.g. "wef"); `postMessage` is the platform-specific
+// statement that ships the {callId, path, args} message to the host.
+inline std::string BuildInitScript(const std::string& ns,
+                                   const std::string& postMessage) {
+  return R"JS(
+(function() {
+  const pendingCalls = new Map();
+  let nextCallId = 1;
+
+  function createWefProxy(path = []) {
+    return new Proxy(function() {}, {
+      get(target, prop) {
+        if (prop === 'then' || prop === 'catch' || prop === 'finally' ||
+            prop === 'constructor' || prop === Symbol.toStringTag) {
+          return undefined;
+        }
+        return createWefProxy([...path, prop]);
+      },
+      apply(target, thisArg, args) {
+        return new Promise((resolve, reject) => {
+          const callId = nextCallId++;
+          pendingCalls.set(callId, { resolve, reject });
+
+          const processedArgs = args.map(arg => {
+            if (typeof arg === 'function') {
+              const cbId = nextCallId++;
+              window.__wefCallbacks = window.__wefCallbacks || {};
+              window.__wefCallbacks[cbId] = arg;
+              return { __callback__: String(cbId) };
+            }
+            if (arg instanceof ArrayBuffer) {
+              const bytes = new Uint8Array(arg);
+              let binary = '';
+              bytes.forEach(b => binary += String.fromCharCode(b));
+              return { __binary__: btoa(binary) };
+            }
+            if (arg instanceof Uint8Array) {
+              let binary = '';
+              arg.forEach(b => binary += String.fromCharCode(b));
+              return { __binary__: btoa(binary) };
+            }
+            return arg;
+          });
+
+          )JS" +
+         postMessage + R"JS(
+        });
+      }
+    });
+  }
+
+  window[")JS" +
+         ns + R"JS("] = createWefProxy();
+
+  window.__wefRespond = function(callId, result, error) {
+    const pending = pendingCalls.get(callId);
+    if (pending) {
+      pendingCalls.delete(callId);
+      if (error) {
+        pending.reject(new Error(error));
+      } else {
+        function convertBinary(obj) {
+          if (obj && typeof obj === 'object') {
+            if (obj.__binary__) {
+              const binary = atob(obj.__binary__);
+              const bytes = new Uint8Array(binary.length);
+              for (let i = 0; i < binary.length; i++) {
+                bytes[i] = binary.charCodeAt(i);
+              }
+              return bytes.buffer;
+            }
+            if (Array.isArray(obj)) {
+              return obj.map(convertBinary);
+            }
+            const result = {};
+            for (const key in obj) {
+              result[key] = convertBinary(obj[key]);
+            }
+            return result;
+          }
+          return obj;
+        }
+        pending.resolve(convertBinary(result));
+      }
+    }
+  };
+
+  window.__wefInvokeCallback = function(callbackId, args) {
+    const cb = window.__wefCallbacks && window.__wefCallbacks[callbackId];
+    if (cb) {
+      cb.apply(null, args);
+    }
+  };
+
+  window.__wefReleaseCallback = function(callbackId) {
+    if (window.__wefCallbacks) {
+      delete window.__wefCallbacks[callbackId];
+    }
+  };
+
+})();
+)JS";
+}
+
+#endif  // WEF_WEBVIEW_INIT_SCRIPT_H_
